Implemented O's and X's on the board.
Implemented O's and X's on the board for 3x3. Rearranging the methods to their respective classes is to be done. Need to work with 6x6 snd 9x9. There is also a lot of code repetation, so new methods are required to minimise the repetation. Only front-end done. Back-end not done yet.